requirements.txt
###### Requirements without Version Specifiers ######`
flask
tweepy
gunicorn
python-dotenv

###### Requirements with Version Specifiers ######`